requirements.txt
future>=0.18.2
pillow>=8.2.0
eel>=0.14.0
matplotlib>=3.4.2
numpy>=1.19.5
deepforest>=1.0.0
tensorflow>=2.5.0